# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BebeMart API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{userId}:
    get:
      operationId: getUserById
      tags: [auth]
      summary: Get user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
...
  /listings/user/{userId}:
    get:
      operationId: getUserListings
      tags: [listings]
      summary: Get listings for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsResponse"
  /transactions:
    get:
      operationId: getMyTransactions
      tags: [transactions]
      summary: Get current user transactions
      responses:
        "200":
          description: Transactions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create transaction (buy now)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{transactionId}:
    get:
      operationId: getTransactionById
      tags: [transactions]
      summary: Get transaction details
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{transactionId}/status:
    patch:
      operationId: updateTransactionStatus
      tags: [transactions]
      summary: Update escrow status
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransactionStatusInput"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
  /reviews/{sellerId}:
    get:
      operationId: getSellerReviews
      tags: [reviews]
      summary: Get reviews for a seller
      parameters:
        - name: sellerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a seller review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [buyer, seller, both]
        state:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
        reviewCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - reviewCount
        - createdAt
    UserPublic:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        state:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
        reviewCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - reviewCount
        - createdAt
    Listing:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: integer
        category:
          type: string
        ageGroup:
          type: string
        condition:
          type: string
        images:
          type: array
          items:
            type: string
        videoUrl:
          type: string
          nullable: true
        sellerId:
          type: integer
        sellerName:
          type: string
        sellerRating:
          type: number
          nullable: true
        state:
          type: string
        city:
          type: string
        status:
          type: string
          enum: [available, sold, reserved]
        giftItForward:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - price
        - category
        - ageGroup
        - condition
        - images
        - sellerId
        - sellerName
        - state
        - city
        - status
        - giftItForward
        - createdAt
...
    ListingsResponse:
      type: object
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - listings
        - total
        - page
        - totalPages
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        listingId:
          type: integer
        listingTitle:
          type: string
        listingImages:
          type: array
          items:
            type: string
        buyerId:
          type: integer
        buyerName:
          type: string
        sellerId:
          type: integer
        sellerName:
          type: string
        amount:
          type: integer
        platformFee:
          type: integer
        escrowStatus:
          type: string
          enum: [pending, paid_escrow, shipped, received, completed, cancelled]
        deliveryAddress:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - listingId
        - listingTitle
        - listingImages
        - buyerId
        - buyerName
        - sellerId
        - sellerName
        - amount
        - platformFee
        - escrowStatus
        - createdAt
        - updatedAt
    CreateTransactionInput:
      type: object
      properties:
        listingId:
          type: integer
        deliveryAddress:
          type: string
      required:
        - listingId
        - deliveryAddress
    UpdateTransactionStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [shipped, received, completed, cancelled]
      required:
        - status
...
    Review:
      type: object
      properties:
        id:
          type: integer
        buyerId:
          type: integer
        buyerName:
          type: string
        sellerId:
          type: integer
        transactionId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - buyerId
        - buyerName
        - sellerId
        - transactionId
        - rating
        - createdAt
    CreateReviewInput:
      type: object
      properties:
        sellerId:
          type: integer
        transactionId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true
      required:
        - sellerId
        - transactionId
        - rating
...