# lib/api-zod/src/generated/types/createTransactionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BebeMart API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTransactionInput {
  listingId: number;
  deliveryAddress: string;
}