# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BebeMart API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  buyerId: number;
  buyerName: string;
  sellerId: number;
  transactionId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string | null;
  createdAt: string;
}